restructured file namings and placement
